fix(relay+runner+web): unify snapshot mechanism and make relay a transparent transport
Consolidates PRs #240, #295, #297 into a single systematic fix for ACP/PTY
snapshot recovery on browser refresh. Previously relay buffered terminal output
and had leaked application logic (Control state machine, input filtering,
app-level Ping/Pong). This removes all non-transport concerns from relay,
unifies the snapshot request mechanism for both PTY and ACP modes, and fixes
the root cause of blank screens on refresh.

Relay:
- Remove output/snapshot ring buffer (-2100 lines) — relay is now a pure
  bidirectional byte pipe between publisher (runner) and subscribers (browsers)
- Remove Control state machine (request/grant/deny/release) — dead code,
  never triggered by any client
- Remove app-level Ping/Pong interception — WebSocket native ping suffices
- Remove ControlRequest type and codec from protocol
- Rename TerminalChannel → Channel (relay is protocol-agnostic)

Runner:
- Rename MsgTypeResync → MsgTypeSnapshotRequest (clear semantics)
- Both PTY and ACP register SnapshotRequest handler (unified mechanism)
- PTY SendSnapshot caches last valid snapshot (survives transient VT clears)
- subscribe_pod only updates token when already connected (no snapshot race)

Web:
- Rename Resync → SnapshotRequest in protocol
- Remove scheduleSnapshotRetry / snapshotReceived / snapshotTimer (no retry needed)
- handleSnapshot: cursor-home only (\x1b[H), no screen clear (graceful size mismatch)
- Remove forceResize from setupConnection (resize driven by xterm onResize)
- globalThis singleton cleanup + beforeunload for HMR stale connection prevention
- Resubscribe sends SnapshotRequest for panel close/reopen recovery
- i18n: add missing zh translations for share dialog
