+++ relay/internal/channel/channel_lifecycle.go
@@ -9,26 +9,22 @@ import (
 	"github.com/anthropics/agentsmesh/relay/internal/protocol"
 )
 
-// AddSubscriber adds a subscriber (browser observer)
-func (c *TerminalChannel) AddSubscriber(subscriberID string, conn *websocket.Conn) {
+func (c *Channel) AddSubscriber(subscriberID string, conn *websocket.Conn) {
 	_ = c.addSubscriberInternal(subscriberID, conn, 0)
 }
 
 // AddSubscriberWithLimit atomically checks subscriber count and adds if under limit.
 // Returns MaxSubscribersError if at capacity. Use maxSubscribers=0 for no limit.
-func (c *TerminalChannel) AddSubscriberWithLimit(subscriberID string, conn *websocket.Conn, maxSubscribers int) error {
+func (c *Channel) AddSubscriberWithLimit(subscriberID string, conn *websocket.Conn, maxSubscribers int) error {
 	return c.addSubscriberInternal(subscriberID, conn, maxSubscribers)
 }
 
-// addSubscriberInternal is the shared implementation for AddSubscriber/AddSubscriberWithLimit.
-func (c *TerminalChannel) addSubscriberInternal(subscriberID string, conn *websocket.Conn, maxSubscribers int) error {
+func (c *Channel) addSubscriberInternal(subscriberID string, conn *websocket.Conn, maxSubscribers int) error {
 	subscriber := &Subscriber{ID: subscriberID, Conn: conn}
 
 	c.subscribersMu.Lock()
 
 	// Check closed INSIDE subscribersMu to prevent TOCTOU race with Close().
-	// Without this, Close() could complete between the closed check and
-	// subscribersMu.Lock(), leaving a goroutine on a dead channel.
 	c.closedMu.RLock()
 	if c.closed {
 		c.closedMu.RUnlock()
@@ -44,7 +40,6 @@ func (c *TerminalChannel) addSubscriberInternal(subscriberID string, conn *webso
 	}
 	c.subscribers[subscriberID] = subscriber
 
-	// Cancel keep-alive timer if exists
 	if c.keepAliveTimer != nil {
 		c.keepAliveTimer.Stop()
 		c.keepAliveTimer = nil
@@ -54,22 +49,6 @@ func (c *TerminalChannel) addSubscriberInternal(subscriberID string, conn *webso
 
 	c.logger.Info("Subscriber connected", "subscriber_id", subscriberID, "total_subscribers", count)
 
-	// Send buffered Output messages to new subscriber
-	// This allows new observers to see recent terminal output they missed
-	bufferedOutput := c.getBufferedOutput()
-	if len(bufferedOutput) > 0 {
-		c.logger.Debug("Sending buffered output to new subscriber",
-			"subscriber_id", subscriberID, "count", len(bufferedOutput))
-		for _, data := range bufferedOutput {
-			if err := subscriber.WriteMessage(data); err != nil {
-				c.logger.Warn("Failed to send buffered output to new subscriber",
-					"subscriber_id", subscriberID, "error", err)
-				break // Stop sending if connection has issues
-			}
-		}
-	}
-
-	// Notify new subscriber if publisher is currently disconnected
 	if c.IsPublisherDisconnected() {
 		if err := subscriber.WriteMessage(protocol.EncodeRunnerDisconnected()); err != nil {
 			c.logger.Warn("Failed to send publisher disconnected status to new subscriber",
@@ -77,13 +56,11 @@ func (c *TerminalChannel) addSubscriberInternal(subscriberID string, conn *webso
 		}
 	}
 
-	// Start forwarding from this subscriber to publisher
 	go c.forwardSubscriberToPublisher(subscriberID)
 	return nil
 }
 
-// RemoveSubscriber removes a subscriber
-func (c *TerminalChannel) RemoveSubscriber(subscriberID string) {
+func (c *Channel) RemoveSubscriber(subscriberID string) {
 	c.subscribersMu.Lock()
 	subscriber, ok := c.subscribers[subscriberID]
 	if !ok {
@@ -97,44 +74,29 @@ func (c *TerminalChannel) RemoveSubscriber(subscriberID string) {
 
 	c.logger.Info("Subscriber disconnected", "subscriber_id", subscriberID, "remaining_subscribers", count)
 
-	// Release control if this subscriber had it
-	c.controllerMu.Lock()
-	if c.controllerID == subscriberID {
-		c.controllerID = ""
-	}
-	c.controllerMu.Unlock()
-
 	if count == 0 {
 		c.handleLastSubscriberGone()
 	}
 }
 
-// handleLastSubscriberGone starts the keep-alive timer when no subscribers remain.
-func (c *TerminalChannel) handleLastSubscriberGone() {
-	// Last subscriber left, start keep-alive timer
+func (c *Channel) handleLastSubscriberGone() {
 	c.subscribersMu.Lock()
 
-	// Double-check: a new subscriber may have connected between the
-	// unlock above and this re-lock, preventing a spurious timer.
 	if len(c.subscribers) != 0 {
 		c.subscribersMu.Unlock()
 		return
 	}
 
-	// Stop any existing timer to prevent duplicates
 	if c.keepAliveTimer != nil {
 		c.keepAliveTimer.Stop()
 	}
 
 	c.lastSubscriberDisconnect = time.Now()
 	c.keepAliveTimer = time.AfterFunc(c.config.KeepAliveDuration, func() {
-		// Guard against race with Close(): if channel is already closed,
-		// skip the callback to avoid spurious onAllSubscribersGone calls.
 		if c.IsClosed() {
 			return
 		}
 
-		// Check if still no subscribers after timeout
 		c.subscribersMu.RLock()
 		stillEmpty := len(c.subscribers) == 0
 		c.subscribersMu.RUnlock()
@@ -149,16 +111,15 @@ func (c *TerminalChannel) handleLastSubscriberGone() {
 	c.subscribersMu.Unlock()
 }
 
-// SubscriberCount returns the number of connected subscribers
-func (c *TerminalChannel) SubscriberCount() int {
+func (c *Channel) SubscriberCount() int {
 	c.subscribersMu.RLock()
 	defer c.subscribersMu.RUnlock()
 	return len(c.subscribers)
 }
 
 // Close closes the channel and all connections.
-// Safe for concurrent callers — only the first call performs cleanup.
-func (c *TerminalChannel) Close() {
+// Safe for concurrent callers -- only the first call performs cleanup.
+func (c *Channel) Close() {
 	c.publisherReplaceMu.Lock()
 	c.closeInternal()
 	c.publisherReplaceMu.Unlock()
@@ -166,7 +127,7 @@ func (c *TerminalChannel) Close() {
 
 // closeInternal performs the actual close logic.
 // MUST be called with publisherReplaceMu held.
-func (c *TerminalChannel) closeInternal() {
+func (c *Channel) closeInternal() {
 	c.closedMu.Lock()
 	if c.closed {
 		c.closedMu.Unlock()
@@ -177,14 +138,12 @@ func (c *TerminalChannel) closeInternal() {
 
 	c.logger.Info("Closing channel")
 
-	// Stop keep-alive timer
 	c.subscribersMu.Lock()
 	if c.keepAliveTimer != nil {
 		c.keepAliveTimer.Stop()
 	}
 	c.subscribersMu.Unlock()
 
-	// Stop publisher reconnect timer and close publisher connection
 	c.publisherMu.Lock()
 	if c.publisherReconnectTimer != nil {
 		c.publisherReconnectTimer.Stop()
@@ -196,11 +155,8 @@ func (c *TerminalChannel) closeInternal() {
 	}
 	c.publisherMu.Unlock()
 
-	// Wait for publisher forwarding goroutine to exit.
-	// conn.Close() above triggers ReadMessage error, causing the goroutine to return.
 	c.publisherWg.Wait()
 
-	// Close all subscriber connections
 	c.subscribersMu.Lock()
 	for _, subscriber := range c.subscribers {
 		_ = subscriber.Conn.Close()
@@ -208,16 +164,6 @@ func (c *TerminalChannel) closeInternal() {
 	c.subscribers = make(map[string]*Subscriber)
 	c.subscribersMu.Unlock()
 
-	// Release output buffer memory
-	c.outputBufferMu.Lock()
-	for i := range c.outputBuffer {
-		c.outputBuffer[i] = nil
-	}
-	c.outputBuffer = nil
-	c.outputBufferBytes = 0
-	c.outputBufferMu.Unlock()
-
-	// Notify channel closed
 	if c.onChannelClosed != nil {
 		c.onChannelClosed(c.PodKey)
 	}
